fix(build): macOS build after local-engine commit
- Cargo.toml: tauri-plugin-dialog, tar, bzip2 and the base llama-cpp-2 entry
  had been swallowed into the [target.'cfg(target_os = "linux")'] table, so
  non-Linux builds lacked the dialog plugin ("Permission dialog:allow-open
  not found"). Move them back to [dependencies].
- system_audio.rs: screencapturekit 1.5 with_sample_rate takes Into<i32>;
  cast the u32 SCK_SAMPLE_RATE.

Verified: `npm run dev` builds and launches on macOS (Apple Silicon).

+++ src-tauri/src/audio/system_audio.rs
@@ -109,7 +109,7 @@ impl SystemAudioCapture {
             .with_height(2)
             .with_captures_audio(true)
             .with_excludes_current_process_audio(true) // Prevent TTS audio feedback loop
-            .with_sample_rate(SCK_SAMPLE_RATE)
+            .with_sample_rate(SCK_SAMPLE_RATE as i32)
             .with_channel_count(2);
 
         // Create channel for audio data
